+++ R/plot-population-time-profile.R
@@ -2,6 +2,10 @@
 #'
 #' @inheritParams plotIndividualTimeProfile
 #' @inheritParams .extractAggregatedSimulatedData
+#' @param showLegendPerDataset Logical flag to display separate legend entries
+#' for observed and simulated datasets, if available. This is experimental and
+#' may not work reliably when both observed and simulated datasets > 1. Defaults
+#' to `FALSE`.
 #' @param ... additionnal arguments to pass to `.extractAggregatedSimulatedData()`
 #'
 #' @inherit .extractAggregatedSimulatedData details
@@ -40,12 +44,14 @@ plotPopulationTimeProfile <- function(dataCombined,
                                       defaultPlotConfiguration = NULL,
                                       aggregation = "quantiles",
                                       quantiles = c(0.05, 0.5, 0.95),
+                                      showLegendPerDataset = FALSE,
                                       ...) {
   .plotTimeProfile(
     dataCombined,
     defaultPlotConfiguration,
     aggregation,
     probs = quantiles,
+    showLegendPerDataset = showLegendPerDataset,
     ...
   )
 }